project/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.1.2",
    "minsearch>=0.0.10",
    "pydantic-ai>=1.73.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
]
